chore: move some components into own files
